Refactor: SQLDriver interface
